+++ src/app/ApplicationTemplate.Shared.Views/Configuration/ViewServicesConfiguration.cs
@@ -19,17 +19,16 @@ public static class ViewServicesConfiguration
 		return services
 			.AddSingleton(s => App.Instance.NavigationMultiFrame.Dispatcher)
 			.AddSingleton(s => Shell.Instance.ExtendedSplashScreen)
-			/*
+
 			.AddSingleton<IDispatcherScheduler>(s => new MainDispatcherScheduler(
 				Shell.Instance.DispatcherQueue,
 				DispatcherQueuePriority.Normal
 			))
-			*/
 			.AddSingleton<IDispatcherFactory, DispatcherFactory>()
 			/*
 			.AddSingleton<IDiagnosticsService, DiagnosticsService>()
 			*/
-			.AddSingleton<ILauncherService>(s => new DispatcherLauncherServiceDecorator(new LauncherService(), App.Instance.Shell.DispatcherQueue))
+			.AddSingleton<ILauncherService>(s => new LauncherService(Shell.Instance.DispatcherQueue))
 			.AddSingleton<IExtendedSplashscreenController, ExtendedSplashscreenController>(s => new ExtendedSplashscreenController(Shell.Instance.DispatcherQueue))
 			.AddSingleton<IConnectivityProvider, ConnectivityProvider>();
 			/*
